test: adjust fixtures, move pegkeeper tests
